==> R/00_run.R <==
#### ATM #######################################################################
#'
#' @name 00_run.R
#'
#' @description Run the entire project.
#'
#' @author Esteban Degetau
#'
#' @created 2024-02-20
#'
#### Run #######################################################################

rm(list = ls())
gc()

#---- Libraries ----------------------------------------------------------------

pacman::p_load(here)

#---- Setup --------------------------------------------------------------------

._01_open_denue <- 1

#---- Run scripts --------------------------------------------------------------

if (._01_open_denue) {
  source(here::here("R", "01_open_denue.R"), encoding = "UTF-8")
}
